Trying out how an atomic structure could work with respect to tables
