+++ src/components/member/tabs/ClaimsTab.tsx
@@ -1,14 +1,12 @@
 import styled from '@emotion/styled'
-import { ClaimSource, useCreateClaimMutation } from 'api/generated/graphql'
-import { ClaimListHeader } from 'components/claims/claims-list/components/ClaimListHeader'
-import { ClaimListItem } from 'components/claims/claims-list/components/ClaimListItem'
-import { format } from 'date-fns'
-import { useGetMemberClaims } from 'graphql/use-get-member-claims'
-import { FadeIn } from 'hedvig-ui/animations/fade-in'
 import {
-  LoadingMessage,
-  StandaloneMessage,
-} from 'hedvig-ui/animations/standalone-message'
+  ClaimSource,
+  GetMemberClaimsDocument,
+  useCreateClaimMutation,
+} from 'api/generated/graphql'
+import { MemberClaimsList } from 'components/organisms/MemberClaimsList'
+import { format } from 'date-fns'
+import { FadeIn } from 'hedvig-ui/animations/fade-in'
 import { Button } from 'hedvig-ui/button'
 import { DateTimePicker } from 'hedvig-ui/date-time-picker'
 import { EnumDropdown } from 'hedvig-ui/dropdown'
@@ -16,7 +14,6 @@ import { Spacing } from 'hedvig-ui/spacing'
 import { MainHeadline } from 'hedvig-ui/typography'
 import React from 'react'
 import { toast } from 'react-hot-toast'
-import { Table } from 'semantic-ui-react'
 
 const HeaderWrapper = styled.div`
   display: flex;
@@ -32,7 +29,6 @@ const FormWrapper = styled.div`
 export const ClaimsTab: React.FC<{
   memberId: string
 }> = ({ memberId }) => {
-  const [claims, { loading, refetch }] = useGetMemberClaims(memberId)
   const [
     createClaim,
     { loading: createClaimLoading },
@@ -41,10 +37,6 @@ export const ClaimsTab: React.FC<{
   const [showForm, setShowForm] = React.useState(false)
   const [claimSource, setClaimSource] = React.useState<ClaimSource | null>(null)
   const [claimDate, setClaimDate] = React.useState<Date>(new Date())
-
-  if (loading || !claims) {
-    return <LoadingMessage paddingTop="25vh" />
-  }
 
   return (
     <FadeIn>
@@ -105,6 +97,12 @@ export const ClaimsTab: React.FC<{
                           date: format(claimDate, "yyyy-MM-dd'T'HH:mm:ss"),
                           source: claimSource,
                         },
+                        refetchQueries: [
+                          {
+                            query: GetMemberClaimsDocument,
+                            variables: { memberId },
+                          },
+                        ],
                       }),
                       {
                         loading: 'Creating claim',
@@ -117,8 +115,6 @@ export const ClaimsTab: React.FC<{
                         error: 'Could not create claim',
                       },
                     )
-
-                    await refetch()
                   }}
                 >
                   Add
@@ -136,20 +132,7 @@ export const ClaimsTab: React.FC<{
           )}
         </FormWrapper>
       </HeaderWrapper>
-      {claims.length === 0 ? (
-        <StandaloneMessage paddingTop="10vh">
-          No claims for member
-        </StandaloneMessage>
-      ) : (
-        <Table celled selectable>
-          <ClaimListHeader />
-          <Table.Body>
-            {claims.map((item, index) => (
-              <ClaimListItem key={item.id} item={item} index={index} />
-            ))}
-          </Table.Body>
-        </Table>
-      )}
+      <MemberClaimsList memberId={memberId} />
     </FadeIn>
   )
 }
